% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aco.R
\name{aco}
\alias{aco}
\title{Ant System Optimization (ACO) algorithm for the (a)symmetric
\href{http://en.wikipedia.org/wiki/Travelling_salesman_problem}{Travelling
Salesperson Problem} (TSP).}
\usage{
aco(x, control, show.info = TRUE)
}
\arguments{
\item{x}{[\code{Network}]\cr
A network, i. e., a graph generated with the \pkg{netgen} package.
See \code{\link[netgen]{makeNetwork}} for details.}

\item{control}{[\code{AntsControl}]\cr
Control object. See \code{\link{makeAntsControl}}.}

\item{show.info}{[\code{logical(1)}]\cr
Should logging messages be printed to the console? Default is \code{FALSE}.}
}
\value{
[\code{AntsResult}]
  S3 result object.
}
\description{
Implementation of the classical Ant System (AS) introduced by Dorigo.
}
\examples{
library(netgen)
  x = generateRandomNetwork(n.points = 6L)
  ctrl = makeAntsControl(
    alpha = 1.2, beta = 1.8, n.ants = 15L,
    init.pher.conc = 0.01, max.iter = 10L
  )
  res = aco(x, ctrl)
  print(res)

  x = matrix(c(1, 2, 1, 3, 1, 4, 3, 1, 3, 2, 3, 4), ncol = 2, byrow = TRUE)
  x = makeNetwork(x, lower = 1, upper = 4)
  ctrl = makeAntsControl(alpha = 1.2, beta = 1.8, n.ants = 20L, max.iter = 30L)
  res = aco(x, ctrl, show.info = TRUE)
}
\keyword{Optimization}

