% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/asNumeric.R
\name{as.numeric.AntsResult}
\alias{as.numeric.AntsResult}
\title{Extracts the best tour found from an AntResult object.}
\usage{
\method{as.numeric}{AntsResult}(x, close = FALSE, ...)
}
\arguments{
\item{x}{[\code{AntResult}]\cr
Result object of a run of the \code{\link{aco}} function.}

\item{close}{[\code{logical(1)}]\cr
Close the tour, i. e., append the last city to the end?}

\item{...}{[any]\cr
Currenly not used.}
}
\description{
Extracts the best tour found from an AntResult object.
}

