% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autoplot.AntsResult.R
\name{autoplot.AntsResult}
\alias{autoplot.AntsResult}
\title{Plots optimization progress.}
\usage{
\method{autoplot}{AntsResult}(object, step.size = 1L, ...)
}
\arguments{
\item{object}{[\code{AntsResult}]\cr
Return value of \code{\link{aco}}.}

\item{step.size}{[\code{integer(1)}]\cr
Which iterations shall be plotted? Default is \code{1L}.}

\item{...}{[any]\cr
Currently not used.}
}
\value{
Nothing.
}
\description{
Plots optimization progress.
}
\note{
This function requires the storage of the entire optimization trace,
  i. e., set the parameter \code{trace.all} of \code{aco} to \code{TRUE}.
}

