% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getTour.R
\name{getTour}
\alias{getTour}
\title{Extracts the best tour found from an \code{AntResult} object.}
\usage{
getTour(x, close = FALSE, ...)
}
\arguments{
\item{x}{[\code{AntResult}]\cr
Result object of a run of the \code{\link{runACOTSP}} function.}

\item{close}{[\code{logical(1)}]\cr
Close the tour, i. e., append the last city to the end? Default is \code{FALSE}.}

\item{...}{[any]\cr
Currenly not used.}
}
\value{
[numeric]
}
\description{
Extracts the best tour found from an \code{AntResult} object.
}

