% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeAntsControl.R
\name{makeAntsControl}
\alias{makeAntsControl}
\title{Generates a control object for the \code{\link{aco}} function.}
\usage{
makeAntsControl(n.ants = 2L, n.elite = n.ants, use.global.best = FALSE,
  best.deposit.only = FALSE, alpha = 1, beta = 2, rho = 0.1,
  att.factor = 1, init.pher.conc = 1e-04, min.pher.conc = 0,
  max.pher.conc = 1e+06, local.search.fun = NULL,
  local.search.step = integer(), prp.prob = 0,
  local.pher.update.fun = NULL, max.iter = 10L, max.time = Inf,
  global.opt.value = NULL, termination.eps = 0.1, trace.all = FALSE)
}
\arguments{
\item{n.ants}{[\code{integer(1)}]\cr
Number of ants. Positive integer.}

\item{n.elite}{[\code{integer(1)}]\cr
Number of \dQuote{elite} ants. Only the \code{n.elite} best ants are allowed to deposit
pheromones during the update phase. Must me lower than or equal to \code{n.ants}.
Default is \code{n.elite = n.ants}.}

\item{use.global.best}{[\code{logical(1)}]\cr
Should the global best tour be used to update the pheromones? Default is \code{FALSE}.
Keep in mind: if \code{use.global.best} is \code{TRUE} and \code{n.elite} is
0, only the global best tour (and consequently only the arcs of the global best
tour) gain pheromones in each iteration.}

\item{best.deposit.only}{[\code{logical(1)}]\cr
Should only the arcs of the global \dQuote{best-so-far} tour be affected by
pheromone evaporation and deposit? Default is \code{FALSE}. This parameter
has an effect if and only if \code{use.global.best} is \code{TRUE}.}

\item{alpha}{[\code{numeric(1)}]\cr
This parameter decides how much influence the pheromones on an edge have
on the selection of edges. Default is \code{1}.}

\item{beta}{[\code{numeric(1)}]\cr
This parameter decided how much influence the edge distances have
on the selection of edges. Default is \code{2}.}

\item{rho}{[\code{numeric(1)}]\cr
\dQuote{Pheromone evaporation coefficient} respectively \dQuote{evaporation rate}.
In each iteration the pheromones on edge (i,j) are first decreased by (1 - rho)
before ants deposit their pheromones on it. Must be in (0, 1). Default is \code{0.1}.}

\item{att.factor}{[\code{numeric(1)}]\cr
This is the socalled \dQuote{constant attractiveness factor}. Default is \code{1}.}

\item{init.pher.conc}{[\code{numeric(1)}]\cr
Initial pheromone concentration for every single edge. Default is \code{0.0001}.}

\item{min.pher.conc}{[\code{numeric(1)}]\cr
Minimal pheromone concentration for every single edge. Default is \code{0}.}

\item{max.pher.conc}{[\code{numeric(1)}]\cr
Maximal pheromone concentration for every single edge. Default is \code{10e5}.}

\item{local.search.fun}{[\code{function}]\cr
Local search procedure which should be applied to all ant tours every
\code{local.search.step} iterations. The function must expect the TSP
instance \code{x} as the first, the initial tour \code{initial.tour} as
the second argument. Default is \code{NULL}, i.e., no local search at all.}

\item{local.search.step}{[\code{integer}]\cr
This parameter determines when to apply the local search procedure given by
the \code{local.search.fun} function. If this is a single scalar value n, the
local search is applied every n iterations. If a integer vector is passed,
the local search procedure is called on exactly the given iterations. This
parameter is considered only if \code{local.search.step} is not \code{NULL}.
Default is the empty vector, which means no local search at all.}

\item{prp.prob}{[\code{numeric(1)}]\cr
Probability used in the pseudo-random-proportional action choice rule used, e.g., by
the Ant Colony System. Default is 0, which means that the rule is not applied
at all and the tour construction is done in the classical way.}

\item{local.pher.update.fun}{[\code{function}]\cr
Local (pheromone) update rule applied right after an ant crossed an edge. Default
is \code{NULL}, which means no local update at all. This must be a function which
expects a single parameter \code{pher}.}

\item{max.iter}{[\code{integer(1)}]\cr
Maximal number of iterations. Default is \code{10}.}

\item{max.time}{[\code{integer(1)}]\cr
Maximum running time in seconds. The algorithm tries hard to hold this
restriction by logging the times of a number of prior iterations and
determining statistically whether the time limit can be hold when another
iteration is done. Default ist \code{Inf}, which means no time limit at all.}

\item{global.opt.value}{[\code{numeric(1)}]\cr
Known global best tour length. This can be used as another termination
criterion. Default is \code{NULL}, which means, that the length of the
globally best tour is unknown.}

\item{termination.eps}{[\code{numeric(1)}]\cr
If \code{global.opt.value} is set, the algorithm stops if the quadratic
distance between the global optimum value and the value of the best tour
found so far is lower than these gap value. Ignored if \code{global.opt.value}
is \code{NULL}.}

\item{trace.all}{[\code{logical(1)}]\cr
Should we save additional information in each iteration, i. e., pheromone
matrix, all ant trails, best ant trail of the current iteration and so on?
Default is \code{FALSE}. You need to set this to \code{TRUE} if you want
to plot the optimization progress via \code{autoplot.AntsResult}.}
}
\value{
[\code{AntsControl}]
  S3 control object containing all the checked parameters and reasonable defaults.
}
\description{
This function generates a control object, i.e., internally a simple list, of
parameters and does sanity checks.
}

