% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotIteration.R
\name{plotIteration}
\alias{plotIteration}
\title{Generates a ggplot object of a single iteration.}
\usage{
plotIteration(network, storage)
}
\arguments{
\item{network}{[\code{Network}]\cr
Network.}

\item{storage}{[\code{list}]\cr
List of internal iteration state containing pher.mat, best.tour and ants.tours
for a single iteration.}
}
\value{
[\code{ggplot}]
}
\description{
Generates a ggplot object of a single iteration.
}

