% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autoplot.AntsResult.R
\name{plotResult}
\alias{plotResult}
\title{Plots optimization progress.}
\usage{
plotResult(object, step.size = 1L, pause = FALSE, ...)
}
\arguments{
\item{object}{[\code{AntsResult}]\cr
Return value of \code{\link{aco}} or one of the shortcut functions, e.g.,
\code{\link{runAS}}.}

\item{step.size}{[\code{integer(1)}]\cr
Which iterations shall be plotted? Default is \code{1L}.}

\item{pause}{[\code{logical(1)}]\cr
Should each plot be displayed after generation? Default is \code{FALSE}.}

\item{...}{[any]\cr
Currently not used.}
}
\value{
[\code{list}] A named list of \code{\link[ggplot2]{ggplot}} objects.
}
\description{
This function generates \code{\link[ggplot2]{ggplot}} objects for each iteration
specified by \code{step.size} and returns these plots in a list. Moreover the
plots can be rendered directly via the option \code{pause}.
}
\note{
This function requires the storage of the entire optimization trace,
  i. e., set the parameter \code{trace.all} of \code{makeAntsControl} to
  \code{TRUE}.
}

