% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{print.AntsResult}
\alias{print.AntsResult}
\title{Prints AntsResult object to standard output.}
\usage{
\method{print}{AntsResult}(x, ...)
}
\arguments{
\item{x}{[\code{AntsResult}]\cr
Result object.}

\item{...}{[any]\cr
Currently not used.}
}
\value{
Nothing
}
\description{
Prints AntsResult object to standard output.
}

