% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runACS.R
\name{runACS}
\alias{runACS}
\title{Wrapper function for the Ant Colony System.}
\usage{
runACS(x, n.ants = 10L, alpha = 1, beta = 2, rho = 0.1,
  att.factor = 1, monitor = makeNullMonitor(), ...)
}
\arguments{
\item{x}{[\code{Network}]\cr
A network, i. e., a graph generated with the \pkg{netgen} package.
See \code{\link[netgen]{makeNetwork}} for details.}

\item{n.ants}{[\code{integer(1)}]\cr
Number of ants. Positive integer.}

\item{alpha}{[\code{numeric(1)}]\cr
This parameter decides how much influence the pheromones on an edge have
on the selection of edges. Default is \code{1}.}

\item{beta}{[\code{numeric(1)}]\cr
This parameter decided how much influence the edge distances have
on the selection of edges. Default is \code{2}.}

\item{rho}{[\code{numeric(1)}]\cr
\dQuote{Pheromone evaporation coefficient} respectively \dQuote{evaporation rate}.
In each iteration the pheromones on edge (i,j) are first decreased by (1 - rho)
before ants deposit their pheromones on it. Must be in (0, 1). Default is \code{0.1}.}

\item{att.factor}{[\code{numeric(1)}]\cr
This is the socalled \dQuote{constant attractiveness factor}. Default is \code{1}.}

\item{monitor}{[\code{AntsMonitor}]\cr
Monitor object of type \code{AntsMonitor}. Determines how to visualize the
optimization process. Default is not visualization. See \code{\link{makeConsoleMonitor}}
for basic console output.}

\item{...}{[\code{any}]\cr
Further parameters passed to control object, e.g., \code{max.iter}. See
\code{makeAntsControl}.}
}
\value{
[\code{AntsResult}]
  S3 result object.
}
\description{
Offers a more R-like interface to the Ant Colony System (ACS) for the TSP.
}

