% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualizePheromoneMatrix.R
\name{visualizePheromoneMatrix}
\alias{visualizePheromoneMatrix}
\title{Helper function to visualize a pheromone matrix.}
\usage{
visualizePheromoneMatrix(pher.mat, ...)
}
\arguments{
\item{pher.mat}{[\code{matrix}]\cr
Square numeric (pheromone) matrix.}

\item{...}{[any]\cr
Currently not used.}
}
\value{
\code{\link[ggplot2]{ggplot}} object.
}
\description{
This function draws a heatmap with darker color indicating higher pheromone
concentration on the corresponding arc.
}

